[2607-DEV-600] Calendar refactor 3: EventPopup decomposition + lazy load
Splits EventPopup.tsx (631 lines) into EventPopupShell (shared header/meta/QR
dialog) plus GuestActions/MemberActions/CoreAdminActions selected by role.
Core has no distinct behavior in the original code, so CoreAdminActions
delegates to MemberActions for the core case rather than duplicating it.
CalendarClient.tsx now loads EventPopup via next/dynamic({ ssr: false }) since
it only renders after a click.
