+++ app/(dashboard)/calendar/components/popup/GuestActions.tsx
@@ -1,0 +1,4 @@
+// Guests get no roles section — mirrors the original `!isGuest && ...` guard in EventPopup.tsx.
+export default function GuestActions() {
+  return null
+}
